Feat: HomeView에HalfASheet구현
