+++ DonWorry/DonWorry/DonWorryApp.swift
@@ -18,8 +18,8 @@ struct DonWorryApp: App {
     var body: some Scene {
         WindowGroup {
 //            ContentView()
-//            HomeView(currentUser: user4)
-            SignInView()
+            HomeView(currentUser: user4)
+//            SignInView()
         }
     }
 }
